+++ pallets/gear/src/manager/mod.rs
@@ -51,12 +51,12 @@ mod task;
 pub use journal::*;
 pub use task::*;
 
-use crate::{Config, CurrencyOf, GearProgramPallet, TaskPoolOf};
+use crate::{Config, CurrencyOf, GasHandlerOf, GearProgramPallet, Pallet, TaskPoolOf};
 use codec::{Decode, Encode};
 use common::{
     event::*,
     scheduler::{ScheduledTask, TaskHandler, TaskPool},
-    ActiveProgram, CodeStorage, Origin, ProgramState,
+    ActiveProgram, CodeStorage, GasTree, Origin, ProgramState,
 };
 use core::fmt;
 use core_processor::common::{Actor, ExecutableActorData};
@@ -64,9 +64,10 @@ use frame_support::traits::Currency;
 use frame_system::pallet_prelude::BlockNumberFor;
 use gear_core::{
     code::{CodeAndId, InstrumentedCode},
-    ids::{CodeId, MessageId, ProgramId},
+    ids::{CodeId, MessageId, ProgramId, ReservationId},
     memory::WasmPageNumber,
     message::{DispatchKind, ExitCode},
+    reservation::GasReservationSlot,
 };
 use primitive_types::H256;
 use sp_runtime::traits::UniqueSaturatedInto;
@@ -278,5 +279,32 @@ where
                 ScheduledTask::RemoveGasReservation(program_id, reservation_id),
             );
         }
+    }
+
+    pub fn remove_gas_reservation(
+        program_id: ProgramId,
+        reservation_id: ReservationId,
+    ) -> GasReservationSlot {
+        let program_id = program_id.into_origin();
+        let mut prog = common::get_active_program(program_id).unwrap_or_else(|e| {
+            unreachable!(
+                "gas reservation removing guaranteed to be called only on existing program: {}",
+                e
+            )
+        });
+        let slot = prog
+                    .gas_reservation_map
+                    .remove(&reservation_id)
+                    .unwrap_or_else(|| unreachable!("gas reservation removing guaranteed to be called only on existing reservation ID"));
+        common::set_program(program_id, prog);
+
+        GasHandlerOf::<T>::unlock_all(reservation_id)
+            .unwrap_or_else(|e| unreachable!("GasTree corrupted! {:?}", e));
+
+        // TODO: charge after unlock (#1830)
+
+        Pallet::<T>::consume_and_retrieve(reservation_id);
+
+        slot
     }
 }
